Handle corrupted registry hives cleanly (#2)
Add additional safeguards when allocating or reading an amount of
data that is dependent on a value from the registry hive.
If the registry hive somehow was corrupted, this value might be
invalid, causing huge allocations or long reads. To limit these
effects, verify in those cases that we read only with our ReaderAt
limits and quit on getting a reader error.